Very basic calender showing the dates and times of "groups" Very early build, still need to add buttons and logic to take to group screen and toggle between create and join
