+++ lib/screens/home_screen.dart
@@ -2,13 +2,7 @@ import 'package:flutter/material.dart';
 import '../utils/colors.dart';
 import '../utils/globals.dart';
 
-class HomeScreen extends StatefulWidget {
-
-  @override
-  _HomeScreenState createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
+class HomeScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
@@ -31,7 +25,36 @@ class _HomeScreenState extends State<HomeScreen> {
                 textAlign: TextAlign.center,
               ),
             ),
-            Home(),
+            Container(
+              width: MediaQuery.of(context).size.width / 1.5,
+              height: MediaQuery.of(context).size.height / 3,
+              margin: const EdgeInsets.symmetric(vertical: 30),
+              child: Image.asset(
+                  'assets/images/johncage.jpg',
+                  fit: BoxFit.fitHeight
+              ),
+            ),
+            Container(
+              width: MediaQuery.of(context).size.width,
+              padding: const EdgeInsets.all(15),
+              child: Row(
+                children: <Widget>[
+                  Flexible(
+                    child: Text(
+                      'In honor of John Cage, this project was made in order to express his quote “Everything we do is music”. '
+                          'You and up to 4 other friends can sign up and schedule a session. From there, everyone in the session can'
+                          'take a recording of any sound and mix them all into one song called a concert. It can be any sound such as '
+                          'waves of a beach, birds chirping, sound of a metro passing by, and so much more.',
+                      style: TextStyle(
+                          fontSize: bioTextSize,
+                          color: textColor
+                      ),
+                      textAlign: TextAlign.center,
+                    ),
+                  ),
+                ],
+              ),
+            ),
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 10),
               child: Text(
@@ -42,7 +65,26 @@ class _HomeScreenState extends State<HomeScreen> {
                 ),
               ),
             ),
-            AboutUs(),
+            Center(
+                child: Container(
+                  width: MediaQuery.of(context).size.width,
+                  padding: const EdgeInsets.all(15),
+                  child: Row(
+                    children: <Widget>[
+                      Flexible(
+                        child: Text(
+                          'We are a group of students from UCF creating this project for our Senior Design Project',
+                          style: TextStyle(
+                              fontSize: bioTextSize,
+                              color: textColor
+                          ),
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                    ],
+                  ),
+                )
+            ),
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 10),
               child: Text(
@@ -60,67 +102,8 @@ class _HomeScreenState extends State<HomeScreen> {
     );
   }
 
-  Widget Home() {
-    return Column(
-      children: <Widget>[
-        Container(
-          width: MediaQuery.of(context).size.width / 1.5,
-          height: MediaQuery.of(context).size.height / 3,
-          margin: const EdgeInsets.symmetric(vertical: 30),
-          child: Image.asset(
-              'assets/images/johncage.jpg',
-              fit: BoxFit.fitHeight
-          ),
-        ),
-        Container(
-          width: MediaQuery.of(context).size.width,
-          padding: const EdgeInsets.all(15),
-          child: Row(
-            children: <Widget>[
-              Flexible(
-                child: Text(
-                  'In honor of John Cage, this project was made in order to express his quote “Everything we do is music”. '
-                      'You and up to 4 other friends can sign up and schedule a session. From there, everyone in the session can'
-                      'take a recording of any sound and mix them all into one song called a concert. It can be any sound such as '
-                      'waves of a beach, birds chirping, sound of a metro passing by, and so much more.',
-                  style: TextStyle(
-                      fontSize: bioTextSize,
-                      color: textColor
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ]
-    );
-  }
-
-  Widget AboutUs() {
-    return Center(
-        child: Container(
-          width: MediaQuery.of(context).size.width,
-          padding: const EdgeInsets.all(15),
-          child: Row(
-            children: <Widget>[
-              Flexible(
-                child: Text(
-                  'We are a group of students from UCF creating this project for our Senior Design Project',
-                  style: TextStyle(
-                      fontSize: bioTextSize,
-                      color: textColor
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-            ],
-          ),
-        )
-    );
-  }
-
   Widget JohnCage() {
     return Container();
   }
 }
+
